completed deleted player functionality. updated console app to support self-hosting for signal R support

+++ services/GraphQL.Api/clientApp/src/components/deleteBtn.js
@@ -6,6 +6,9 @@ import './deleteBtn.css';
 export const DeleteBtn = (props) => (
     <button className="btn btn-danger p-del" onClick={e => {
         e.preventDefault();
-        props.onDelete();
+        if(window.confirm(props.message)) 
+        {
+            props.onDelete();
+        }
     }}>X</button>
 );
